// src/front/src/api/cdr/models/callRecordResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Cdr API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export interface CallRecordResponse {
	/** @nullable */
	gid?: string | null;
	/** @nullable */
	callerId?: string | null;
	/** @nullable */
	calledNumber?: string | null;
	startTime?: string;
	endTime?: string;
	duration?: number;
	billableSeconds?: number;
	/** @nullable */
	callStatusCode?: string | null;
	/** @nullable */
	callStatusName?: string | null;
	/** @nullable */
	callDirectionCode?: string | null;
	totalCost?: number;
	/** @nullable */
	currencyCode?: string | null;
	/** @nullable */
	customerName?: string | null;
	createdAt?: string;
}